+++ ving/cli/jobs.mjs
@@ -1,5 +1,6 @@
 import { defineCommand, showUsage } from "citty";
 import { VingJobWorker } from '#ving/jobs/worker.mjs';
+import { pause, resume, drain } from '#ving/jobs/queue.mjs';
 import { generateJobHandler } from '#ving/generator/jobhandler.mjs';
 import ving from '#ving/index.mjs';
 
@@ -9,18 +10,36 @@ export default defineCommand({
         description: "Manage background jobs",
     },
     args: {
-        worker: {
-            type: "boolean",
-            description: "Start a worker",
-            alias: "w",
-            default: false,
-        },
         queueName: {
             type: "string",
             description: "Set a queue name. Defaults to `jobs`.",
             valueHint: 'jobs',
             default: 'jobs',
             alias: 'q',
+        },
+        drain: {
+            type: "boolean",
+            description: "Delete all waiting and delayed jobs in the queue.",
+            alias: "D",
+            default: false,
+        },
+        pause: {
+            type: "boolean",
+            description: "Stop workers from executing jobs in the queue.",
+            alias: "P",
+            default: false,
+        },
+        resume: {
+            type: "boolean",
+            description: "Resume workers executing jobs in the queue.",
+            alias: "R",
+            default: false,
+        },
+        worker: {
+            type: "boolean",
+            description: "Start a worker",
+            alias: "w",
+            default: false,
         },
         ttl: {
             type: "number",
@@ -38,7 +57,31 @@ export default defineCommand({
             type: "string",
             description: "Specify a JSON string of data you'd like to pass into the job.",
             default: '{}',
+            alias: 'j',
+        },
+        cron: {
+            type: "string",
+            description: "Specify a 5 parameter cron string for how often the new job should run.",
+            valueHint: '* * * * * ',
+            alias: 'c',
+        },
+        repeat: {
+            type: "number",
+            description: "Specify a number of milliseconds to wait before this job is executed again.",
+            valueHint: 30000,
+            alias: 'r',
+        },
+        delay: {
+            type: "number",
+            description: "The number of milliseconds to wait before executing this job.",
+            valueHint: 30000,
             alias: 'd',
+        },
+        priority: {
+            type: "number",
+            description: "A number ranging from `1` to `2097152` where `1` is the highest possible priority. Defaults to `2097152`.",
+            valueHint: 100,
+            alias: 'p',
         },
         handler: {
             type: 'string',
@@ -49,7 +92,19 @@ export default defineCommand({
     },
     async run({ args, cmd }) {
         try {
-            if (args.worker) {
+            if (args.drain) {
+                await drain({ queueName: args.queueName });
+                ving.close();
+            }
+            else if (args.pause) {
+                await pause({ queueName: args.queueName });
+                ving.close();
+            }
+            else if (args.resume) {
+                await resume({ queueName: args.queueName });
+                ving.close();
+            }
+            else if (args.worker) {
                 const worker = new VingJobWorker(args.queueName);
                 await worker.start();
                 if (args.ttl > 0) {
@@ -63,9 +118,16 @@ export default defineCommand({
                 ving.close();
             }
             else if (args.addJob) {
-                await ving.addJob(args.addJob, JSON.parse(args.jobData), {
-                    queueName: args.queueName,
-                });
+                const params = { queueName: args.queueName };
+                if (args.cron)
+                    params.cron = args.cron;
+                if (args.repeat)
+                    params.repeat = args.repeat;
+                if (args.priority)
+                    params.priority = args.priority;
+                if (args.delay)
+                    params.delay = args.delay;
+                await ving.addJob(args.addJob, JSON.parse(args.jobData), params);
                 if (!args.worker)
                     ving.close();
             }